Create Find Customer Referee
To retrieve the names of all customers who were not referred by the customer with ID = 2, we need to consider cases where referee_id is either not equal to 2 or is NULL. Since NULL values indicate no referee, they should also be included in the output. Therefore, the condition should ensure that either referee_id is not equal to 2 or it is NULL, allowing all qualifying records to be displayed.