Create Big Countries
To find the big countries, we need to check if a country meets at least one of two conditions: having a large area or a large population. A country is considered big if its total land area is 3 million square kilometers or more, or if its population is at least 25 million people.

To get this information, we look at the data and select only those countries that satisfy either of these conditions. If a country meets at least one of them, it will be included in the final list. The result will display the name of each qualifying country along with its population and area.