Create Article Views 1
To find all authors who have viewed at least one of their own articles, we need to check cases where the author of an article and the viewer of that article are the same person. This means the author_id and viewer_id must be equal.

Since there can be duplicate records in the table, we ensure that each author appears only once in the results by using the ""Distinct" Keyword. After identifying the authors who meet this condition, we return their IDs in ascending order for a structured and clear output using "Order By"