Create Invalid Tweets
To find the IDs of invalid tweets, we use the LENGTH() function to determine the number of characters in the content of each tweet. The LENGTH() function counts the total number of characters in the tweet's content. If this number is greater than 15, the tweet is considered invalid.

By checking this condition for each tweet, we can identify the tweet_id of all tweets that have content exceeding 15 characters and mark them as invalid. The result will return the IDs of all such invalid tweets.