Create 1378. Replace Employee ID With The Unique Identifier
We need to display the unique ID of each user, ensuring that if a user does not have a unique ID, it appears as NULL. Since we require all records from the Employee table (the left table), we can use a LEFT JOIN with the EmployeeUNI table. This will include all rows from the Employee table, displaying NULL for unmatched records while showing the corresponding unique ID for matching ones.