Create 1068. Product Sales Analysis I
We have two tables, Sales and Product, where the product_id column in the Sales table acts as a foreign key referencing the Product table. To display the product_name, year, and price for each product_id in the Sales table, we need to join both tables on the common product_id column. Since we want to retrieve all the details of product_id present in the Sales table (the left table), we will use a LEFT JOIN to combine the two tables.